PI3 (2)/PI3 (1)/PI3/requirements.txt
# BOSS SHOPP Database Requirements
# Dependências Python para o sistema de banco de dados

# Banco de dados
mysql-connector-python==8.2.0

# Segurança
bcrypt==4.1.2

# Utilitários
python-decouple==3.8

# Desenvolvimento e testes
pytest==7.4.3
pytest-cov==4.1.0

# Logging e monitoramento
colorlog==6.8.0

# Data e tempo
python-dateutil==2.8.2

# Validação de dados
cerberus==1.3.5

# Backup e compressão
pymysql==1.1.0

# CLI e interface
click==8.1.7
rich==13.7.0

# Configuração
pyyaml==6.0.1

# Desenvolvimento
black==23.11.0
flake8==6.1.0
mypy==1.7.1